Services/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
h11==0.16.0
httptools==0.6.4
idna==3.10
loguru==0.7.3
pydantic==2.11.7
pydantic-core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing-extensions==4.15.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
win32_setctime==1.2.0
typer[all]
httpx
readability-lxml
trafilatura
PyMuPDF
python-docx
beautifulsoup4
langdetect
openai
pydantic-settings
mammoth 
tiktoken
blingfire
numpy
